'代码优化'

+++ src/main/java/io/renren/modules/sys/dao/SysMenuDao.java
@@ -24,7 +24,7 @@ public interface SysMenuDao extends BaseMapper<SysMenuEntity> {
 	 * 根据父菜单，查询子菜单
 	 * @param parentId 父菜单ID
 	 */
-	List<SysMenuEntity> queryListParentId(Long parentId);
+	List<SysMenuEntity> queryListParentId(Integer parentId);
 	
 	/**
 	 * 获取不包含按钮的菜单列表
