+++ src/main/java/io/renren/modules/sys/dao/SysRoleDao.java
@@ -23,5 +23,5 @@ public interface SysRoleDao extends BaseMapper<SysRoleEntity> {
 	/**
 	 * 查询用户创建的角色ID列表
 	 */
-	List<Long> queryRoleIdList(Long createUserId);
+	List<Long> queryRoleIdList(Integer createUserId);
 }
