+++ src/main/java/io/renren/modules/sys/dao/SysRoleMenuDao.java
@@ -23,10 +23,10 @@ public interface SysRoleMenuDao extends BaseMapper<SysRoleMenuEntity> {
 	/**
 	 * 根据角色ID，获取菜单ID列表
 	 */
-	List<Long> queryMenuIdList(Long roleId);
+	List<Integer> queryMenuIdList(Integer roleId);
 
 	/**
 	 * 根据角色ID数组，批量删除
 	 */
-	int deleteBatch(Long[] roleIds);
+	int deleteBatch(Integer[] roleIds);
 }
