+++ src/main/java/io/renren/modules/sys/dao/SysUserDao.java
@@ -24,12 +24,12 @@ public interface SysUserDao extends BaseMapper<SysUserEntity> {
 	 * 查询用户的所有权限
 	 * @param userId  用户ID
 	 */
-	List<String> queryAllPerms(Long userId);
+	List<String> queryAllPerms(Integer userId);
 	
 	/**
 	 * 查询用户的所有菜单ID
 	 */
-	List<Long> queryAllMenuId(Long userId);
+	List<Integer> queryAllMenuId(Integer userId);
 	
 	/**
 	 * 根据用户名，查询系统用户
