+++ src/main/java/io/renren/modules/sys/dao/SysUserRoleDao.java
@@ -23,11 +23,11 @@ public interface SysUserRoleDao extends BaseMapper<SysUserRoleEntity> {
 	/**
 	 * 根据用户ID，获取角色ID列表
 	 */
-	List<Long> queryRoleIdList(Long userId);
+	List<Integer> queryRoleIdList(Long userId);
 
 
 	/**
 	 * 根据角色ID数组，批量删除
 	 */
-	int deleteBatch(Long[] roleIds);
+	int deleteBatch(Integer[] roleIds);
 }
