+++ src/main/java/io/renren/modules/sys/service/ShiroService.java
@@ -21,11 +21,11 @@ public interface ShiroService {
     /**
      * 获取用户权限列表
      */
-    Set<String> getUserPermissions(long userId);
+    Set<String> getUserPermissions(Integer userId);
 
     /**
      * 根据用户ID，查询用户
      * @param userId
      */
-    SysUserEntity queryUser(Long userId);
+    SysUserEntity queryUser(Integer userId);
 }
