+++ src/main/java/io/renren/modules/sys/service/SysMenuService.java
@@ -27,13 +27,13 @@ public interface SysMenuService extends IService<SysMenuEntity> {
 	 * @param parentId 父菜单ID
 	 * @param menuIdList  用户菜单ID
 	 */
-	List<SysMenuEntity> queryListParentId(Long parentId, List<Long> menuIdList);
+	List<SysMenuEntity> queryListParentId(Integer parentId, List<Integer> menuIdList);
 
 	/**
 	 * 根据父菜单，查询子菜单
 	 * @param parentId 父菜单ID
 	 */
-	List<SysMenuEntity> queryListParentId(Long parentId);
+	List<SysMenuEntity> queryListParentId(Integer parentId);
 	
 	/**
 	 * 获取不包含按钮的菜单列表
@@ -43,10 +43,10 @@ public interface SysMenuService extends IService<SysMenuEntity> {
 	/**
 	 * 获取用户菜单列表
 	 */
-	List<SysMenuEntity> getUserMenuList(Long userId);
+	List<SysMenuEntity> getUserMenuList(Integer userId);
 
 	/**
 	 * 删除
 	 */
-	void delete(Long menuId);
+	void delete(Integer menuId);
 }
