+++ src/main/java/io/renren/modules/sys/service/SysRoleMenuService.java
@@ -22,16 +22,16 @@ import java.util.List;
  */
 public interface SysRoleMenuService extends IService<SysRoleMenuEntity> {
 	
-	void saveOrUpdate(Long roleId, List<Long> menuIdList);
+	void saveOrUpdate(Integer roleId, List<Integer> menuIdList);
 	
 	/**
 	 * 根据角色ID，获取菜单ID列表
 	 */
-	List<Long> queryMenuIdList(Long roleId);
+	List<Integer> queryMenuIdList(Integer roleId);
 
 	/**
 	 * 根据角色ID数组，批量删除
 	 */
-	int deleteBatch(Long[] roleIds);
+	int deleteBatch(Integer[] roleIds);
 	
 }
