+++ src/main/java/io/renren/modules/sys/service/SysRoleService.java
@@ -29,11 +29,11 @@ public interface SysRoleService extends IService<SysRoleEntity> {
 
 	void update(SysRoleEntity role);
 
-	void deleteBatch(Long[] roleIds);
+	void deleteBatch(Integer[] roleIds);
 
 	
 	/**
 	 * 查询用户创建的角色ID列表
 	 */
-	List<Long> queryRoleIdList(Long createUserId);
+	List<Long> queryRoleIdList(Integer createUserId);
 }
