+++ src/main/java/io/renren/modules/sys/service/SysUserRoleService.java
@@ -22,15 +22,15 @@ import java.util.List;
  */
 public interface SysUserRoleService extends IService<SysUserRoleEntity> {
 	
-	void saveOrUpdate(Long userId, List<Long> roleIdList);
+	void saveOrUpdate(Integer userId, List<Integer> roleIdList);
 	
 	/**
 	 * 根据用户ID，获取角色ID列表
 	 */
-	List<Long> queryRoleIdList(Long userId);
+	List<Integer> queryRoleIdList(Long userId);
 
 	/**
 	 * 根据角色ID数组，批量删除
 	 */
-	int deleteBatch(Long[] roleIds);
+	int deleteBatch(Integer[] roleIds);
 }
