+++ src/main/java/io/renren/modules/sys/service/SysUserService.java
@@ -29,12 +29,12 @@ public interface SysUserService extends IService<SysUserEntity> {
 	 * 查询用户的所有权限
 	 * @param userId  用户ID
 	 */
-	List<String> queryAllPerms(Long userId);
+	List<String> queryAllPerms(Integer userId);
 	
 	/**
 	 * 查询用户的所有菜单ID
 	 */
-	List<Long> queryAllMenuId(Long userId);
+	List<Integer> queryAllMenuId(Integer userId);
 
 	/**
 	 * 根据用户名，查询系统用户
@@ -62,5 +62,5 @@ public interface SysUserService extends IService<SysUserEntity> {
 	 * @param password     原密码
 	 * @param newPassword  新密码
 	 */
-	boolean updatePassword(Long userId, String password, String newPassword);
+	boolean updatePassword(Integer userId, String password, String newPassword);
 }
