+++ src/main/java/io/renren/modules/sys/service/impl/SysRoleMenuServiceImpl.java
@@ -29,16 +29,16 @@ public class SysRoleMenuServiceImpl extends ServiceImpl<SysRoleMenuDao, SysRoleM
 
 	@Override
 	@Transactional(rollbackFor = Exception.class)
-	public void saveOrUpdate(Long roleId, List<Long> menuIdList) {
+	public void saveOrUpdate(Integer roleId, List<Integer> menuIdList) {
 		//先删除角色与菜单关系
-		deleteBatch(new Long[]{roleId});
+		deleteBatch(new Integer[]{roleId});
 
 		if(menuIdList.size() == 0){
 			return ;
 		}
 
 		//保存角色与菜单关系
-		for(Long menuId : menuIdList){
+		for(Integer menuId : menuIdList){
 			SysRoleMenuEntity sysRoleMenuEntity = new SysRoleMenuEntity();
 			sysRoleMenuEntity.setMenuId(menuId);
 			sysRoleMenuEntity.setRoleId(roleId);
@@ -48,12 +48,12 @@ public class SysRoleMenuServiceImpl extends ServiceImpl<SysRoleMenuDao, SysRoleM
 	}
 
 	@Override
-	public List<Long> queryMenuIdList(Long roleId) {
+	public List<Integer> queryMenuIdList(Integer roleId) {
 		return baseMapper.queryMenuIdList(roleId);
 	}
 
 	@Override
-	public int deleteBatch(Long[] roleIds){
+	public int deleteBatch(Integer[] roleIds){
 		return baseMapper.deleteBatch(roleIds);
 	}
